consolidated _props.go files
