--- pkg/api/latest/dynakube/logmonitoring_props.go
@@ -1,19 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package dynakube
-
-import (
-	"github.com/Dynatrace/dynatrace-operator/pkg/api/latest/dynakube/logmonitoring"
-)
-
-func (dk *DynaKube) LogMonitoring() *logmonitoring.LogMonitoring {
-	lm := &logmonitoring.LogMonitoring{
-		Spec:         dk.Spec.LogMonitoring,
-		TemplateSpec: dk.Spec.Templates.LogMonitoring,
-	}
-	lm.SetName(dk.Name)
-	lm.SetHostAgentDependency(dk.OneAgent().IsDaemonsetRequired())
-
-	return lm
-}
