--- pkg/api/latest/dynakube/metadata_enrichment_props.go
@@ -1,13 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package dynakube
-
-import "github.com/Dynatrace/dynatrace-operator/pkg/api/latest/dynakube/metadataenrichment"
-
-func (dk *DynaKube) MetadataEnrichment() *metadataenrichment.MetadataEnrichment {
-	return &metadataenrichment.MetadataEnrichment{
-		Spec:   &dk.Spec.MetadataEnrichment,
-		Status: &dk.Status.MetadataEnrichment,
-	}
-}
