--- pkg/api/latest/dynakube/oneagent_props.go
@@ -1,23 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package dynakube
-
-import (
-	"github.com/Dynatrace/dynatrace-operator/pkg/api/latest/dynakube/oneagent"
-)
-
-func (dk *DynaKube) OneAgent() *oneagent.OneAgent {
-	oa := oneagent.NewOneAgent(
-		&dk.Spec.OneAgent,
-		&dk.Status.OneAgent,
-		&dk.Status.CodeModules,
-		dk.Name,
-		dk.APIURLHost(),
-		dk.FF().IsOneAgentPrivileged(),
-		dk.FF().SkipOneAgentLivenessProbe(),
-		dk.GetResourceAttributes(),
-	)
-
-	return oa
-}
