--- pkg/api/latest/dynakube/telemetryservice_props.go
@@ -1,17 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package dynakube
-
-import (
-	"github.com/Dynatrace/dynatrace-operator/pkg/api/latest/dynakube/telemetryingest"
-)
-
-func (dk *DynaKube) TelemetryIngest() *telemetryingest.TelemetryIngest {
-	ts := &telemetryingest.TelemetryIngest{
-		Spec: dk.Spec.TelemetryIngest,
-	}
-	ts.SetName(dk.Name)
-
-	return ts
-}
